Update to make ui alert across pages
