+++ airflow-core/src/airflow/ui/src/pages/Dashboard/Dashboard.tsx
@@ -18,57 +18,30 @@
  */
 import { Box, Heading, VStack } from "@chakra-ui/react";
 
-import type { UIAlert } from "openapi/requests/types.gen";
-import ReactMarkdown from "src/components/ReactMarkdown";
-import { Accordion, Alert } from "src/components/ui";
-import { useConfig } from "src/queries/useConfig";
+import { UIAlertAccordion } from "src/components/UIAlertAccordion";
 
 import { Health } from "./Health";
 import { HistoricalMetrics } from "./HistoricalMetrics";
 import { PoolSummary } from "./PoolSummary";
 import { Stats } from "./Stats";
 
-export const Dashboard = () => {
-  const alerts = useConfig("dashboard_alert") as Array<UIAlert>;
-
-  return (
-    <Box overflow="auto" px={4}>
-      <VStack alignItems="start">
-        {alerts.length > 0 ? (
-          <Accordion.Root collapsible defaultValue={["ui_alerts"]}>
-            <Accordion.Item key="ui_alerts" value="ui_alerts">
-              {alerts.map((alert: UIAlert, index) =>
-                index === 0 ? (
-                  <Accordion.ItemTrigger key={alert.text} mb={2}>
-                    <Alert status={alert.category}>
-                      <ReactMarkdown>{alert.text}</ReactMarkdown>
-                    </Alert>
-                  </Accordion.ItemTrigger>
-                ) : (
-                  <Accordion.ItemContent key={alert.text} pr={8}>
-                    <Alert status={alert.category}>
-                      <ReactMarkdown>{alert.text}</ReactMarkdown>
-                    </Alert>
-                  </Accordion.ItemContent>
-                ),
-              )}
-            </Accordion.Item>
-          </Accordion.Root>
-        ) : undefined}
-        <Heading mb={2} size="2xl">
-          Welcome
-        </Heading>
-      </VStack>
-      <Box>
-        <Stats />
-      </Box>
-      <Box display="flex" gap={8} mt={8}>
-        <Health />
-        <PoolSummary />
-      </Box>
-      <Box mt={8}>
-        <HistoricalMetrics />
-      </Box>
+export const Dashboard = () => (
+  <Box overflow="auto" px={4}>
+    <VStack alignItems="start">
+      <UIAlertAccordion />
+      <Heading mb={2} size="2xl">
+        Welcome
+      </Heading>
+    </VStack>
+    <Box>
+      <Stats />
     </Box>
-  );
-};
+    <Box display="flex" gap={8} mt={8}>
+      <Health />
+      <PoolSummary />
+    </Box>
+    <Box mt={8}>
+      <HistoricalMetrics />
+    </Box>
+  </Box>
+);
